Changed exercise repo header to filter button
